requirements.txt
PyYAML>=6.0
PyMuPDF>=1.24.0
google-genai>=1.0.0
